config: add GCM support
Enable GCM ang GHASH to make ipsec working on MBX T1 boxes.

On these boxes, the CPU is not supporting the AES-NI instruction set, so we
need to do it by software.  As ipsec relies on the kernel crypto, we need to
add the GCM support to the kernel.

References: jira PAM-4176